Move Application BOM route into EE
